requirements-scraper.txt
# Requirements for Westpac Scraper
playwright>=1.49.0
python-dotenv>=1.0.0
